Callisto Scam Alert
⛔️⛔️⛔️⛔️⛔️

Documents should be published so that everyone is informed and do not fall into the trap of team fraudsters

❌❌❌❌❌

Post it so that fewer people fall into the trap of these fraudsters. This project is completely fraudulent

🆘🆘🆘🆘🆘

The market size of this coin can be destroyed with only $ 100. In all exchange offices, robots are buying and selling and inviting people to keep the coins for a long time if they sell all the coins themselves.

🌐🌐🌐🌐🌐

Publish

@callistonet
@callistonetnews
Callisto.network